Easy Programming

+++ CUED with function based/djangorest/drapi/views.py
@@ -1,36 +1,13 @@
-"""
 from django.shortcuts import render
 from .models import Aiquest
 from .serializer import AiquestSerializer
-from rest_framework.views import APIView
+from rest_framework.decorators import api_view
 from rest_framework.response import Response
-"""
-from .models import Aiquest
-from .serializer import AiquestSerializer
-from rest_framework.generics import GenericAPIView
-from rest_framework.mixins import ListModelMixin,CreateModelMixin
 
 
-class AiquestList(GenericAPIView,ListModelMixin):
-    queryset=Aiquest.objects.all()
-    serializer_class=AiquestSerializer
-    
-    def get(self,request,*args, **kwargs):
-        return self.list(request,*args,**kwargs)
-    
-
-class AiquestCreate(GenericAPIView,CreateModelMixin):
-    queryset=Aiquest.objects.all()
-    serializer_class=AiquestSerializer
-    
-    def post(self,request,*args, **kwargs):
-        return self.create(request,*args,**kwargs)
-
-
-
-"""
-class AiquestCreate(APIView):  
-    def get(self,request,pk=None,format=None):
+@api_view(['GET','POST','PUT','PATCH','DELETE'])
+def aiquest_create(request,pk=None):
+    if request.method=='GET':
         id=pk
         if id is not None:
             #complex data
@@ -45,14 +22,14 @@ class AiquestCreate(APIView):
         serializer=AiquestSerializer(ai,many=True)
         return Response(serializer.data)
         
-    def post(self,request,format=None):
+    if request.method=='POST':
         serializer=AiquestSerializer(data=request.data)
         if serializer.is_valid():
             serializer.save()
             return Response({'msg':'Successfully insert data'})
         return Response(serializer.errors)
 
-    def put(self,request,pk,format=None):
+    if request.method == 'PUT':
         id=pk
         ai=Aiquest.objects.get(pk=id)
         serializer=AiquestSerializer(ai,data=request.data)
@@ -62,7 +39,7 @@ class AiquestCreate(APIView):
         return Response(serializer.errors)
     
     
-    def patch(self,request,pk,format=None):
+    if request.method == 'PATCH':
         id=pk
         ai=Aiquest.objects.get(pk=id)
         serializer=AiquestSerializer(ai,data=request.data,partial=True)
@@ -72,9 +49,12 @@ class AiquestCreate(APIView):
         return Response(serializer.errors)
 
 
-    def delete(self,request,pk,format=None):
+    if request.method=='DELETE':
         id=pk
         ai=Aiquest.objects.get(pk=id)
         ai.delete()
         return Response({'msg':'Successfully deleted data'})
-"""        
+        
+
+
+
